requirements.txt
numpy==2.3.2
pandas==2.2.2
yfinance==0.2.65
matplotlib==3.8.4
flask==3.0.0
openai==1.54.0
schedule==1.2.0
python-dotenv==1.0.0
email-validator==2.1.0
jinja2==3.1.2
pandas-datareader==0.10.0
requests==2.31.0
python-dateutil==2.8.2
